feat(app): merge billing into the Settings screen
Billing was its own owner-only destination duplicating the account-settings
audience. Fold it in as a "Plan & billing" section on Settings so owners manage
their plan alongside business details and the danger zone.

- Add the plan/seats summary card to settings.tsx, loading billing on mount
  for owners (reusing the shared messageFor error helper).
- Drop the Billing nav item from the sidebar / bottom-tab navigation.
- Redirect the /billing route to /settings so existing links and bookmarks
  still resolve.
